feat: 추천 태그 리스트 및 커스텀 버튼 컴포넌트 추가, 추천 폼 로직 개선
